+++ app/src/main/java/com/jukco/waitforme/ui/sign/social/NaverAuthUiProvider.kt
@@ -2,7 +2,7 @@ package com.jukco.waitforme.ui.sign.social
 
 import android.content.Context
 import android.util.Log
-import com.jukco.waitforme.ui.sign.sign_up.SignUpForm
+import com.jukco.waitforme.ui.sign.sign_up.SignUpDto
 import com.navercorp.nid.NaverIdLoginSDK
 import com.navercorp.nid.oauth.NidOAuthLogin
 import com.navercorp.nid.oauth.OAuthLoginCallback
@@ -14,7 +14,7 @@ import kotlin.coroutines.resume
 class NaverAuthUiProvider(
     private val activityContext: Context,
 ) : AuthUiProvider {
-    override suspend fun signIn(): SignUpForm? =
+    override suspend fun signIn(): SignUpDto? =
         if (signInAndCheckSuccess()) {
             getUserInfo()
         } else {
@@ -41,12 +41,12 @@ class NaverAuthUiProvider(
         NaverIdLoginSDK.authenticate(activityContext, oauthLoginCallback)
     }
 
-    private suspend fun getUserInfo(): SignUpForm? = suspendCancellableCoroutine { continuation ->
+    private suspend fun getUserInfo(): SignUpDto? = suspendCancellableCoroutine { continuation ->
         val nidProfileCallback = object : NidProfileCallback<NidProfileResponse> {
             override fun onSuccess(response: NidProfileResponse) {
                 val profile = response.profile
                 if (profile != null) {
-                    continuation.resume(SignUpForm(profile))
+                    continuation.resume(SignUpDto(profile))
                 } else {
                     Log.e(TAG, "사용자 정보 요청 실패")
                     continuation.resume(null)
